Fix the watch folder logic for CSS.
The initial passthrough would work fine, but afterwards, the
directory structure was lost. By manually copying the style.css
file and listing the sub-directories, the structure is preserved
both on a full build and on a development incremental build.

This problem goes away once we have an asset pipeline through
something like vite.