Tag partials and layouts inside the html
This simple change allows you to quickly see and find partials and
layouts in rendered HTML. This helps to improve the developer experience
by making it possible to quickly see if the system is rendering what you
expect it to render (for example, the right layout) and to do a search
on data-partial="some-string" which is unique to find the correct file
to edit.

This does not affect us negatively. The amount of bytes sent down the
wire is negligible.