Simplifies the language switcher to always lead to the home page
As this is a hard to solve problem with our overall site structure, especially when tags are involved, we, as in the bestuur, agreed on simplifying the switcher to always point to the home page, as otherwise this holds us back from launching. Being sent to the main page seemed an acceptable compromise between bugfreeness/maintainability and desired functionality.